requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiomysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart